refactor: Client Interceptor 클래스 분리
